Updated decoration to support S4 generic methods (closes #84, #11)
Only body of a function is decorated. Packages are loaded before is decorated so
any `setClass`, `setMethod` can run before the decoration proceed. This is
important since otherwise S3 generic methods that will be converted in to S4
generic methods will be broken. Good example is the jsonlite package and the
`plot` method. Simplified the decoration API.